feat(sops): add secrets via sops(-nix)
Add secrets via [sops][sops] and [sops-nix][sops-nix] in order to be
able to publicly store sensitive information in the public repo. The
current key used is an [age][age] key, as it sufficed for the time
being. `sops-nix` imports the host SSH key as well, but this is no yet
used.

Secrets are stored and imported from the `secrets/` directory, while the
`sops` config is found under `.sops.yaml`.

[age]: https://github.com/FiloSottile/age
[sops]: https://github.com/getsops/sops
[sops-nix]: https://github.com/Mic92/sops-nix

Signed-off-by: Filip Dutescu <filip@hucksy.dev>